Remove wrong version of CCT from Model Folder
