+++ Tests/Models/CCT/CCT/registry.py
