+++ Tests/Models/CCT/CCT/utils/helpers.py
