+++ Tests/Models/CCT/CCT/utils/stochastic_depth.py
