--- Tests/Models/CCT/__init__.py
@@ -1,3 +1,0 @@
-from .cct import *
-from .cvt import *
-from .vit import *
