+++ Tests/Models/CCT/mnistCheckpoint.py
