--- Tests/Models/CCT/registry.py
@@ -1,7 +1,0 @@
-
-
-def register_model(func):
-    """
-    Fallback wrapper in case timm isn't installed
-    """
-    return func
