--- Tests/Models/CCT/text/__init__.py
@@ -1,3 +1,0 @@
-from .vit import *
-from .cvt import *
-from .cct import *
