+++ Tests/Models/CCT/utils/utils.py
